requirements.txt
autots
black
darts
fastapi
isort
joblib
matplotlib
numpy < 2.0
pandas < 2.0
pylint
pymongo
ruff
scikit-learn
sort-requirements
torch
torchvision
pytorch-tsmixer
uvicorn
xgboost
